+++ src/main/gram.c
@@ -4932,7 +4932,7 @@ int isValidName(const char *name)
     {
         /* the only way to establish which chars are alpha etc is to
            use the wchar variants */
-        int n = strlen(name), used;
+        size_t n = strlen(name), used;
         wchar_t wc;
         used = Mbrtowc(&wc, p, n, NULL);
         p += used;
